+++ src/Modules/Development/FeatureToggles/Enums/FeatureToggleDefinitions.php
@@ -13,6 +13,7 @@ enum FeatureToggleDefinitions: string
 {
     case ACHIEVEMENT_SYSTEM = 'achievementSystem';
     case MAIL_SEARCH = 'mailSearch';
+    case HYGIENE_QUIZ = 'hygieneQuiz';
 
     /**
      * Returns all feature toggle identifiers.